# swagger.yaml
swagger: "2.0"
info:
  description: "This is a food delivery service app for restaurants. You can find out more at [Fast-Food-Fast](https://fast-food-fast-delivery.herokuapp.com/)."
  version: "1.0.0"
  title: "Fast-Food-Fast"
  termsOfService: "fast-food-fast-delivery.herokuapp.com/terms/"
  contact:
    email: "jideajayi11@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "fast-food-fast-delivery.herokuapp.com"
basePath: "/api/v1"
...
schemes:
- "https"
- "http"
paths:
...
  /menu:
    post:
      tags:
      - "food"
      summary: "Admin can add food to menu"
      description: ""
      operationId: "addMenu"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Food details object"
        required: true
        schema:
          $ref: "#/definitions/food"
      responses:
        400:
          description: "Incomplete body parameters"
        201:
          description: "Food Added"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
    get:
      tags:
      - "food"
      summary: "List Menu "
      description: ""
      operationId: "getMenu"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: ""
        description: ""
        required: false
        schema:
          $ref: "#/definitions/food"
      responses:
        400:
          description: "Incomplete body parameters"
        201:
          description: "Food Added"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
...